requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
azure-ai-textanalytics==5.3.0
    # via news-aggregator (pyproject.toml)
azure-common==1.1.28
    # via azure-ai-textanalytics
azure-core==1.34.0
    # via
    #   azure-ai-textanalytics
    #   azure-cosmos
azure-cosmos==4.9.0
    # via news-aggregator (pyproject.toml)
azure-functions==1.23.0
    # via news-aggregator (pyproject.toml)
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
dotenv==0.9.9
    # via news-aggregator (pyproject.toml)
idna==3.10
    # via requests
isodate==0.7.2
    # via azure-ai-textanalytics
markupsafe==3.0.2
    # via werkzeug
newsapi-python==0.2.7
    # via news-aggregator (pyproject.toml)
python-dotenv==1.1.1
    # via dotenv
requests==2.32.4
    # via
    #   azure-core
    #   newsapi-python
six==1.17.0
    # via azure-core
typing-extensions==4.14.0
    # via
    #   azure-ai-textanalytics
    #   azure-core
    #   azure-cosmos
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via azure-functions
